IEMEREDU13-60(user-pets): Add basic authorization

+++ src/main/java/com/leverx/service/impl/PetServiceImpl.java
@@ -3,18 +3,15 @@ package com.leverx.service.impl;
 import static com.leverx.dto.converter.PetConverterDto.convertPetEntityToResponse;
 import static com.leverx.dto.converter.PetConverterDto.convertListOfEntityToListOfResponse;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
+
+import javax.persistence.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.leverx.dto.response.PetResponseDto;
-import com.leverx.entity.EPetType;
-import com.leverx.exception.NoSuchPetTypeException;
-import com.leverx.exception.PetNotFoundException;
 import com.leverx.repository.PetRepository;
 import com.leverx.service.PetService;
 
@@ -31,10 +28,11 @@ public class PetServiceImpl implements PetService {
 
   @Override
   @Transactional
-  public Optional<PetResponseDto> findById(final long id) {
-    return Optional.of(
-        convertPetEntityToResponse(
-            petRepository.findById(id).orElseThrow(PetNotFoundException::new)));
+  public PetResponseDto findById(final long id) {
+    return convertPetEntityToResponse(
+        petRepository
+            .findById(id)
+            .orElseThrow(() -> new EntityNotFoundException("There is no such pet")));
   }
 
   @Override
@@ -45,23 +43,10 @@ public class PetServiceImpl implements PetService {
 
   @Override
   @Transactional
-  public List<PetResponseDto> getAllByOwnerId(final long ownerId) {
-    return convertListOfEntityToListOfResponse(petRepository.findAllByOwnerId(ownerId));
-  }
-
-  @Override
-  @Transactional
-  public List<PetResponseDto> findAllByOwnerIdAndPetType(final long id, final String type) {
-    if (Arrays.stream(EPetType.values()).anyMatch(t -> t.toString().equals(type.toUpperCase()))) {
-      return convertListOfEntityToListOfResponse(
-              petRepository.findAllByOwnerIdAndPetType(id, EPetType.valueOf(type.toUpperCase())));
+  public void delete(final long id) {
+    if (!petRepository.existsById(id)) {
+      throw new EntityNotFoundException("There is no such pet");
     }
-    throw new NoSuchPetTypeException();
-  }
-
-  @Override
-  @Transactional
-  public void delete(long id) {
     petRepository.deleteById(id);
   }
 }
